backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.ini
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

aiomysql==0.2.0
    # via -r requirements.ini
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via starlette
click==8.1.7
    # via
    #   -r requirements.ini
    #   uvicorn
colorama==0.4.6
    # via click
dependency-injector==4.44.0
    # via -r requirements.ini
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.6
    # via -r requirements.ini
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
pydantic==2.10.4
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.0
    # via -r requirements.ini
pyjwt==2.10.1
    # via -r requirements.ini
pymysql==1.1.1
    # via aiomysql
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.19
    # via -r requirements.ini
six==1.17.0
    # via dependency-injector
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.ini
starlette==0.41.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.33.0
    # via -r requirements.ini
